Almost match DisplaySprite

+++ src/data/sprite_tables.c
@@ -2,7 +2,7 @@
 
 extern const u8 gObjTiles_4bpp[];
 extern const u8 gObjTiles_8bpp[];
-extern const u8 gUnknown_08358528[];
+extern const u32 gUnknown_08358528[];
 
 const struct SpriteTables gSpriteTables = {
      gAnimations,
